[backend] Interlinking Articles with Rating

+++ starter-project-backend-g32/src/controllers/article.ts
@@ -2,6 +2,8 @@ import { Request, Response } from 'express'
 import Article from '../models/article'
 import Rating  from '../models/rating'
 
+const stars_arr = ["one", "two", "three", "four", "five"];
+
 export const getSpecificArticle = async (req: Request, res: Response) => {
         try {
             const doc = await Article.findOne({ _id: req.params.articleID }).lean().exec()
@@ -75,10 +77,7 @@ export const deleteArticle = async (req: Request, res: Response) => {
         }
   }
 
-  // Rating CRUD: 
-//   router.route("/:articleID/rating").post(createRating).get(getAllRating)
-// router.route("/:articleID/rating/:ratingId").get(getSpecificRating).put(updateRating).delete(deleteRating)
-
+//Controller for Rating
 export const getSpecificRating = async (req: Request, res: Response) => {
     try {
         const doc = await Rating.findOne({ _id: req.params.ratingId, articleId: req.params.articleID}).lean().exec()
@@ -93,6 +92,11 @@ export const getSpecificRating = async (req: Request, res: Response) => {
 
 export const getAllRating = async (req: Request, res: Response) => {
     try {
+        const theArticle = await Article.findOne({ _id: req.params.articleID }).lean().exec()
+        if(!theArticle){
+            console.log("No Article by that ID to getALL: ", theArticle)
+            return res.status(404).end()
+        }
         const docs = await Rating.find({ articleId: req.params.articleID }).lean().exec()
         res.status(200).json({ data: docs })
     } catch (e) {
@@ -100,7 +104,7 @@ export const getAllRating = async (req: Request, res: Response) => {
     }
 }
 
-//Handle Interlinking rating and Article
+
 export const createRating = async (req: Request, res: Response) => {
     try {
         //check if user already rated
@@ -110,11 +114,36 @@ export const createRating = async (req: Request, res: Response) => {
             console.log("User already Rated: ", userRated)
             res.status(404).end()
         }else{
+            const theArticle = await Article.findOne({ _id: req.params.articleID }).lean().exec()
+            if(!theArticle){
+                console.log("No Article by that ID: ", theArticle)
+                return res.status(404).end()
+            }
+            //Now try to create the Rating and it will only proceed if there are no errors
             const doc = await Rating.create({
                 stars: req.body.stars,
                 articleId: req.params.articleID,
                 userId: req.body.userId
             })
+            //If you succeed creating the Rating then Update the respective article rating accordingly
+            const stars = stars_arr[req.body.stars -1];
+            const articleRating = theArticle.rating;
+            if(articleRating)
+            { 
+                articleRating[stars] +=1; 
+                await Article.findOneAndUpdate(
+                    {
+                        _id: req.params.articleID,
+                    },
+                    {
+                        rating: {...articleRating}
+                    }
+                    ,
+                    { new: true }
+                    ).lean().exec()
+
+            }else{console.log("Article rating not updated when creating a rating: ", theArticle)}
+
             res.status(201).json({ data: doc  })
         }
     } catch (e) {
@@ -122,9 +151,28 @@ export const createRating = async (req: Request, res: Response) => {
     }
 }
 
-// Interlink with Article
+
 export const updateRating = async (req: Request, res: Response) => {
     try {
+        //first check if the article exists 
+        const theArticle = await Article.findOne({ _id: req.params.articleID }).lean().exec()
+        if(!theArticle){
+            console.log("No Article by that ID: to UpdateRating", theArticle)
+            return res.status(404).end()
+        }
+
+        //if the article exists then decrement the old value of stars in rating and increment the new one
+        //Find the article's rating first
+        const theRating = await Rating.findOne({ _id: req.params.ratingId, articleId: req.params.articleID}).lean().exec()
+        if(!theRating){
+            console.log("No Rating by that ID: to UpdateRating")
+            return res.status(404).end()
+        }
+        //Reserve the old number to decrement it in articles
+        const oldStarNumber = theRating.stars;
+        const oldStar = stars_arr[Number(oldStarNumber) - 1]
+
+        //try updating the new value on the Rating first || it maybe subject to errors so better try first
         const updatedDoc = await Rating
             .findOneAndUpdate(
             { 
@@ -139,17 +187,57 @@ export const updateRating = async (req: Request, res: Response) => {
             .lean()
             .exec()
         if (!updatedDoc) {
-            return res.status(500).end()
+            return res.status(404).end()
         }
+
+        //Now update the article rating accordingly
+        const articleRating = theArticle.rating;
+        const stars = stars_arr[req.body.stars -1];
+        if(articleRating)
+        { 
+            articleRating[oldStar] -=1;
+            articleRating[stars] +=1;
+            await Article.findOneAndUpdate(
+                {
+                    _id: req.params.articleID,
+                },
+                {
+                    rating: {...articleRating}
+                }
+                ,
+                { new: true }
+                ).lean().exec()
+
+        }else{console.log("Article rating not updated when Updating a rating: ", theArticle)}
+
+
         res.status(200).json({ data: updatedDoc })
     } catch (e) {
         res.status(500).end()
     }
 }
 
-//Interlink with Article
+
 export const deleteRating = async (req: Request, res: Response) => {
     try {
+        //first check if the article exists 
+        const theArticle = await Article.findOne({ _id: req.params.articleID }).lean().exec()
+        if(!theArticle){
+            console.log("No Article by that ID: to UpdateRating", theArticle)
+            return res.status(404).end()
+        }
+
+        //Find the article's rating to decrement it's value in articles
+        const theRating = await Rating.findOne({ _id: req.params.ratingId, articleId: req.params.articleID}).lean().exec()
+        if(!theRating){
+            console.log("No Rating by that ID: to UpdateRating")
+            return res.status(404).end()
+        }
+        //Reserve the old number to decrement it in articles
+        const oldStarNumber = theRating.stars;
+        const oldStar = stars_arr[Number(oldStarNumber) - 1]
+
+        //Remove the rating || must be done first as it maybe subjected to errors
         const removed = await Rating.findOneAndRemove(
         { 
             _id: req.params.ratingId,
@@ -158,6 +246,24 @@ export const deleteRating = async (req: Request, res: Response) => {
         if (!removed) {
             return res.status(404).end()
         }
+
+        //after removing successfully update the article accordingly
+        const articleRating = theArticle.rating;
+        if(articleRating)
+        { 
+            articleRating[oldStar] -=1;
+            await Article.findOneAndUpdate(
+                {
+                    _id: req.params.articleID,
+                },
+                {
+                    rating: {...articleRating}
+                }
+                ,
+                { new: true }
+                ).lean().exec()
+
+        }else{console.log("Article rating not updated when Deleting a rating: ", theArticle)}
         return res.status(200).json({ data: removed })
     } catch (e) {
         res.status(400).end()
